Change custom user model app name
The name is being changed from 'auth' to 'accounts' in order to
distinguish the app name from Django's built-in 'auth' app. It is
ambiguous in some circumstances (like settings files) which one is
being referenced.

This is a precursor to potentially changing and using Django's built-in
user model instead of our custom user model.

Change-Id: Ibef5734ace55f38e2fc9d98bf6870e7036ad9a9f